Add myelin-areas bundle. Add room script for look which in retrospect seems like a bit much but it does cut down on the look command by a lot.Make it so that looking at an entity emits a look command on it.
