+++ bundles/myelin-areas/areas/limbo/items.yml
@@ -1,0 +1,185 @@
+- id: 1
+  name: "Rusty Sword"
+  type: WEAPON
+  roomDesc: "Rusted Sword"
+  keywords: [ "sword", "rusty", "metal", "rusted" ]
+  description: "An arm's-length, jagged metal sword discolored with red corrosion. The worn leather grip barely held on by fraying thread."
+  script: "1-sword"
+  metadata:
+    level: 1
+    slot: 'wield'
+    itemLevel: 1
+    quality: common
+    minDamage: 7
+    maxDamage: 15
+    speed: 2.8
+    stats:
+      critical: 1
+- id: 2
+  name: "Slice of Cheese"
+  roomDesc: "A moldy slice of cheese"
+  keywords: [ "slice", "cheese", "moldy" ]
+  description: "A yellow, slightly moldy slice of cheese. Only a rat could find this appetizing."
+  behaviors:
+    decay:
+      duration: 240
+- id: 3
+  type: CONTAINER
+  name: "Wooden Chest"
+  roomDesc: "A wooden chest rests in the corner, its hinges badly rusted."
+  keywords: [ "wooden", "chest" ]
+  description: "Time has not been kind to this chest. It seems to be held together solely by the dirt and rust."
+  items: [ "limbo:1", "limbo:6", "limbo:7", "limbo:8" ]
+  closed: true
+  maxItems: 5
+  metadata:
+    noPickup: true
+- id: 4
+  name: "Scraps"
+  quality: poor
+  roomDesc: "Scraps from a Training Dummy"
+  keywords: [ "dummy", "scraps" ]
+  description: "Splintered, shattered, and generally destroyed remains of a training dummy"
+  behaviors:
+    sellable:
+      value: 5
+      currency: gold
+- id: 5
+  name: "Training Sword"
+  type: WEAPON
+  roomDesc: "Training Sword"
+  keywords: ["sword", "training", "iron"]
+  description: "A hefty iron blade. Not the sharpest sword in the world but it will get the job done."
+  metadata:
+    quality: rare
+    slot: "wield"
+    itemLevel: 10
+    level: 5
+    minDamage: 11
+    maxDamage: 20
+    speed: 2.8
+    stats:
+      strength: 2
+      stamina: 2
+      critical: -1
+  behaviors:
+    sellable:
+      value: 30
+      currency: gold
+- id: 6
+  name: "Leather Vest"
+  type: ARMOR
+  roomDesc: "Leather Vest"
+  keywords: ["leather", "vest"]
+  description: "A plain leather vest. Better than nothing."
+  metadata:
+    quality: common
+    slot: "chest"
+    level: 1
+    itemLevel: 1
+    stats:
+      armor: 20
+  behaviors:
+    sellable:
+      value: 30
+      currency: gold
+- id: 7
+  name: "Potion of Health I"
+  roomDesc: "Potion of Health I"
+  keywords: [ "potion", "health" ]
+  type: POTION
+  metadata:
+    level: 1
+  behaviors:
+    # usable has two variants: spell, and effect
+    usable:
+      # this is the spell variant, you specify a skill of type SPELL
+      spell: "potion"
+      # charges/destroyOnDepleted lets you create consumable items
+      # without specifying charges the item can be used forever
+      charges: 5
+      destroyOnDepleted: true
+      # cooldown before item can be used again
+      cooldown: 30
+      # spell configuration, spell can access it via this.options
+      options:
+        restores: 30
+        stat: "health"
+- id: 8
+  name: "Potion of Strength I"
+  roomDesc: "Potion of Strength I"
+  keywords: [ "potion", "strength" ]
+  type: POTION
+  metadata:
+    level: 1
+  behaviors:
+    usable:
+      # this is the effect variant, you specify an effect name instead of a spell
+      effect: "potion.buff"
+      charges: 2
+      destroyOnDepleted: true
+      # effect config, see Effect docs. `name` is auto-populated with the item name
+      config:
+        description: "Increases strength by <b>10</b> for <b>15</b> seconds"
+        duration: 15000
+      # effect state, see Effect docs
+      state:
+        magnitude: 10
+        stat: "strength"
+- id: 9
+  name: "Blade of Ranvier"
+  type: WEAPON
+  roomDesc: "Blade of Ranvier"
+  keywords: ["sword", "blade", "ranvier"]
+  description: "The blade shines a brilliant silver. Holding it you feel as if you could take on the world."
+  script: "9-ranvier-blade"
+  metadata:
+    quality: epic
+    slot: "wield"
+    itemLevel: 15
+    level: 10
+    minDamage: 13
+    maxDamage: 26
+    speed: 2.8
+    specialEffects:
+      - "Chance on hit: Blade of Ranvier thirsts for blood and heals the wielder for 25% of damage done."
+    stats:
+      strength: 2
+      stamina: 2
+      critical: 3
+- id: 10
+  name: "Wooden Shield"
+  type: ARMOR
+  roomDesc: "Wooden Shield"
+  keywords: ["shield", "wooden"]
+  description: "A rather uninteresting looking wooden shield. A rusted metal band barely hold its together and the leather arm band is nearly torn."
+  metadata:
+    quality: common
+    slot: "shield"
+    level: 1
+    itemLevel: 1
+    stats:
+      armor: 10
+  behaviors:
+    sellable:
+      value: 30
+      currency: gold
+- id: test_key
+  name: "Oddly-shaped Key"
+  keywords: ["key", "odd", "oddly", "shaped"]
+  roomDesc: "A strange looking key"
+  description: "This key seems overly complex with numerous grooves."
+  metadata:
+    quality: common
+- id: locked_chest
+  type: CONTAINER
+  name: "Locked Chest"
+  roomDesc: "A wooden chest rests open in the corner, its hinges badly rusted."
+  keywords: [ "locked", "wooden", "chest" ]
+  items: [ "limbo:1" ]
+  closed: true
+  locked: true
+  lockedBy: "limbo:test_key"
+  maxItems: 5
+  metadata:
+    noPickup: true
