+++ bundles/myelin-areas/areas/limbo/loot-pools.yml
@@ -1,0 +1,6 @@
+---
+potions:
+  - "limbo:7": 10
+  - "limbo:8": 5
+junk:
+  - "limbo:4": 50
