+++ bundles/myelin-areas/areas/limbo/manifest.yml
@@ -1,0 +1,3 @@
+title: "Limbo"
+info:
+  respawnInterval: 30
