+++ bundles/myelin-areas/areas/limbo/npcs.yml
@@ -1,0 +1,141 @@
+- id: 1
+  keywords: ['rat']
+  name: A Rat
+  level: 2
+  description: "The rat's beady red eyes dart frantically, its mouth foaming as it scampers about."
+  script: '1-rat'
+  items: ['limbo:2']
+  quests: ['limbo:2']
+  behaviors:
+    combat: true
+    ranvier-wander:
+      interval: 20
+      restrictTo: ['limbo:1', 'limbo:2', 'limbo:3']
+  attributes:
+    health: 100
+    energy: 100
+- id: 2
+  keywords: ["wise", "old", "man"]
+  name: "Wise Old Man"
+  level: 99
+  script: '2-old-man'
+  pacifist: true
+  description: "A wise looking old man sits on the ground with legs crossed."
+  behaviors:
+    ranvier-sentient: true
+- id: 3
+  keywords: ["puppy", "dog", "loyal", "wide", "eyed", "wide-eyed"]
+  name: A Puppy
+  description: "A wide-eyed puppy stares up at you."
+  pacifist: true
+  level: 1
+  script: '3-puppy'
+- id: 4
+  keywords: ["dummy", "target", "practice"]
+  level: 2
+  name: "Training Dummy"
+  attributes:
+    health: 100
+    strength: 10
+  behaviors:
+    combat: true
+    lootable:
+      # currencies is a list of currencies this npc can possibly drop and the amount range
+      # the currency key is arbitrary so if you want a new currency just add a new key
+      # the key must be formatted_like_this though, for rendering/saving purposes
+      currencies:
+        gold:
+          min: 10
+          max: 20
+      # pools is a list of either a `"itemEntityReference": dropChance` or a name of a loot pool entry
+      # defined in the loot-pools.yml. This lets you do things like having a 'global' loot pool that every
+      # npc in the area can drop
+      pools:
+        - "limbo:junk"
+        - "limbo:potions"
+        - "limbo:2": 25
+  description: >-
+    The training dummy is almost human shaped although slightly out of proportion. The material it's made of is hard to
+    discern; it seems to constantly change between metal, wood, cloth, and glass depending on the angle. There is a
+    large red and white bullseye painted on its chest. The dummy has no eyes and mindlessly meanders about the area.
+- id: 5
+  keywords: ["boss", "target", "dummy", "practice"]
+  name: "Boss Training Dummy"
+  level: 5
+  description: >-
+    This dummy is significantly larger than the others. Bright red with a monstrous figure it lumbers around the area
+    with a great echoing stomp. Where the other target dummies have a bullseye this dummy has a yellow exclamation mark.
+  attributes:
+    health: 200
+    strength: 15
+  behaviors:
+    combat: true
+    lootable:
+      currencies:
+        gold:
+          min: 50
+          max: 100
+      pools:
+        - "limbo:potions"
+        - "limbo:5": 100
+        - "limbo:9": 5
+- id: 6
+  keywords: ["wally", "wonderful", "shop", "vendor"]
+  name: "Wally the Wonderful"
+  description: >-
+    Moe's Shop has the best wares in town! Armor, weapons and potions, you name and we ... might have it!
+  level: 99
+  pacifist: true
+  behaviors:
+    ranvier-sentient: true
+    vendor:
+      items:
+        "limbo:5":
+          cost: 30
+          currency: gold
+        "limbo:6":
+          cost: 30
+          currency: gold
+        "limbo:10":
+          cost: 30
+          currency: gold
+        "limbo:7":
+          cost: 100
+          currency: gold
+        "limbo:8":
+          cost: 150
+          currency: gold
+        "limbo:9":
+          cost: 99999
+          currency: gold
+      enterMessage: "Step right up! Get your wares at Moe's Shop!"
+      leaveMessage: "Come back soon!"
+- id: aggro-player-test
+  keywords: ["test", "aggro", "dummy"]
+  level: 2
+  name: "Player-aggressive Training Dummy"
+  description: >-
+    This NPC is aggressive towards players but not other NPCs. Be careful.
+  attributes:
+    health: 120
+    strength: 12
+  behaviors:
+    combat: true
+    ranvier-aggro:
+      delay: 5
+- id: aggro-npc-test
+  keywords: ["test", "aggro", "dummy"]
+  level: 2
+  name: "Self-hating Training Dummy"
+  description: >-
+    This NPC is aggressive towards other NPCs but not to the player.
+  attributes:
+    health: 100
+    strength: 15
+  behaviors:
+    combat: true
+    ranvier-aggro:
+      delay: 5
+      towards:
+        players: false
+        npcs: ["limbo:aggro-npc-test"]
