+++ bundles/myelin-areas/areas/limbo/quests.yml
@@ -1,0 +1,86 @@
+---
+- id: 1
+  title: "A Journey Begins"
+  level: 1
+  autoComplete: true
+  description: |-
+    A voice whispers to you: Welcome to the world, young one. This is a dangerous and deadly place, you should arm yourself.
+     - Open the chest with '<white>open chest</white>'
+     - Use '<white>get sword chest</white>' and '<white>get vest chest</white>' to get some gear
+     - Equip it using '<white>wield sword</white>' and '<white>wear vest</white>'
+  completionMessage: |-
+    <b><cyan>Hint: You can use the '<white>tnl</white>' or '<white>level</white>' commands to see how much experience you need to level.</cyan></b>
+    <b><yellow>The rat looks like it is hungry, use '<white>quest list rat</white>' to see what aid you can offer. Use '<white>quest start rat 1</white>' to accept their task.</yellow></b>
+    <b><cyan>Hint: To move around the game type any of the exit names listed in <white>[Exits: ...]</white> when you use the '<white>look</white>' command.</cyan></b>
+  goals:
+    - type: FetchGoal
+      config:
+        title: Find a Weapon
+        count: 1
+        item: 'limbo:1'
+    - type: FetchGoal
+      config:
+        title: Find Some Armor
+        count: 1
+        item: 'limbo:6'
+    - type: EquipGoal
+      config:
+        title: Wield A Weapon
+        slot: wield
+    - type: EquipGoal
+      config:
+        title: Equip Some Armor
+        slot: chest
+  rewards:
+    - type: ExperienceReward
+      config:
+        amount: 5
+        leveledTo: QUEST
+    - type: CurrencyReward
+      config:
+        currency: gold
+        amount: 10
+- id: 2
+  title: One Cheese Please
+  level: 1
+  repeatable: true
+  description: |-
+    A rat's squeaks seem to indicate it wants some cheese. You check around the area, maybe someone has left some lying around.
+    
+    Once you find some bring it back to the rat, use '<white>quest log</white>' to find the quest number, then complete the quest with '<white>quest complete #</white>'
+  completionMessage: |-
+    <b><cyan>Hint: NPCs with quests available have <white>[</white><yellow>!</yellow><white>]</white> in front of their name, <white>[</white><yellow>?</yellow><white>]</white> means you have a quest ready to turn in, and <white>[</white><yellow>%</yellow><white>]</white> means you have a quest in progress.</cyan>
+  goals:
+    - type: FetchGoal
+      config:
+        title: Found Cheese
+        count: 1
+        item: 'limbo:2'
+        removeItem: true
+  rewards:
+    - type: ExperienceReward
+      config:
+        amount: 3
+        leveledTo: QUEST
+- id: 3
+  title: Self Defense 101
+  level: 2
+  requires: [ 'limbo:1' ]
+  autoComplete: true
+  description: |-
+    A voice whispers to you: It would be wise to practice protecting yourself. There are a number of training dummies in this area that, while not pushovers, will not be too difficult.
+    - Use '<white>attack dummy</white>' to start combat against the training dummy
+    - Once it's dead any loot it drops will be in its corpse on the ground. You can use '<white>look in corpse</white>' to check again or '<white>loot corpse</white>' to retrieve all your loot.
+  completionMessage: |-
+    <b><cyan>Hint: You can get the loot from enemies with '<white>get <item> corpse</white>' but be quick about it, the corpse will decay after some time.</cyan>
+  goals:
+    - type: KillGoal
+      config:
+        title: Kill a Training Dummy
+        npc: 'limbo:4'
+        count: 1
+  rewards:
+    - type: ExperienceReward
+      config:
+        amount: 5
+        leveledTo: QUEST
