+++ bundles/myelin-areas/areas/limbo/rooms.yml
@@ -1,0 +1,169 @@
+- id: 1
+  title: Test Room 1
+  description: "A featureless white room. A pitch black void in the shape of archway can be seen on the east side of the room."
+  npcs: ["limbo:1" ]
+  items:
+    - id: "limbo:3"
+      respawnChance: 20
+      replaceOnRespawn: true # when this chest respawns completely replace it so its contents get refreshed
+  script: "1-test"
+  exits:
+    - roomId: "limbo:2"
+      direction: east
+      leaveMessage: " steps into the void and disappears."
+    - roomId: "limbo:8"
+      direction: down
+    - roomId: "limbo:9"
+      direction: west
+    - roomId: "mapped:start"
+      direction: north
+- id: 2
+  title: "Test Room 2"
+  description: >-
+    A completely black room. Somehow all of the light that should be coming from the room to the west does not pass
+    through the archway. A single lightbulb hangs from the ceiling illuminating a small area. To the east you see a
+    large white dome. There is a sign above the entrance to the dome: "Training Area"
+  script: "2-look"
+  items:
+    - id: "limbo:2"
+      respawnChance: 10
+  npcs: ["limbo:2", "limbo:3"]
+  exits:
+    - roomId: "limbo:1"
+      direction: west
+      leaveMessage: " steps into the light and disappears."
+    - roomId: "limbo:3"
+      direction: east
+- &trainingroom
+  id: 3
+  title: "Training Room"
+  script: "combat-training"
+  description: >-
+    The entire area is covered by a large dome with a hexagonal grid surface. A beautiful blue sky reaches from horizon
+    to horizon, punctuated by the lines of the grid. The dome shimmers as virtual birds fly into and out of its surface.
+    The pure green grass is eerily undisturbed by you walking over it or by the simulated breeze.
+  npcs:
+    - id: "limbo:4"
+      respawnChance: 25
+      maxLoad: 3
+  exits:
+    - roomId: "limbo:2"
+      direction: west
+    - roomId: "limbo:4"
+      direction: north
+    - roomId: "limbo:6"
+      direction: east
+- <<: *trainingroom
+  id: 4
+  items:
+    - id: "craft:1"
+      respawnChance: 30
+  npcs:
+    - id: "limbo:4"
+      respawnChance: 25
+      maxLoad: 3
+    - id: "limbo:aggro-player-test"
+      respawnChance: 25
+      maxLoad: 1
+  exits:
+    - roomId: "limbo:3"
+      direction: south
+    - roomId: "limbo:5"
+      direction: east
+- <<: *trainingroom
+  id: 5
+  items:
+    - id: "craft:2"
+      respawnChance: 15
+  exits:
+    - roomId: "limbo:4"
+      direction: west
+    - roomId: "limbo:6"
+      direction: south
+    - roomId: "limbo:7"
+      direction: north
+- <<: *trainingroom
+  id: 6
+  npcs:
+    - id: "limbo:4"
+      respawnChance: 25
+      maxLoad: 3
+    - id: "limbo:aggro-npc-test"
+      respawnChance: 50
+      maxLoad: 2
+  exits:
+    - roomId: "limbo:3"
+      direction: west
+    - roomId: "limbo:5"
+      direction: north
+- id: 7
+  title: "Boss Training Room"
+  description: >-
+    The dome in this section is bright red, the pure green grass is replaced with a smooth white surface. The ground
+    beneath your feet has the word "Danger" in bright red letters tiled across the area.
+  npcs:
+    - id: "limbo:5"
+      respawnChance: 50
+  exits:
+    - roomId: "limbo:5"
+      direction: south
+- id: 8
+  title: "Ancient Wayshrine (Waypoint)"
+  description: >
+    A runed black obelisk towers in the center of this clearing, surrounded by a faerie ring. The runes pulse and glow with
+    a soft blue light. The grass immediately around the obelisk is immaculate in stark contrast to the dying former meadow
+    that makes up the clearing. The obelisk looks like it can be used as a <b><cyan>waypoint</cyan></b>.
+  script: '8-waypoint'
+  behaviors:
+    waypoint: true
+  exits:
+    - roomId: "limbo:1"
+      direction: up
+    - roomId: "limbo:10"
+      direction: down
+- id: 9
+  title: "Wally's Wonderful Wares (Shop)"
+  description: >- 
+    A very brightly colored shop stall stands in the middle of an otherwise desolate clearing. The stall is covered in
+    colorful cloth, shining gems, and battle gear of all varieties. A large sign sits next to the products:
+    "<b><yellow>Wally's Wonderful Wares</yellow></b> has the best products in town! Armor, weapons and potions, you name
+    and we... might have it!"
+  npcs:
+    - id: "limbo:6"
+      respawnChance: 0
+  exits:
+    - roomId: "limbo:1"
+      direction: east
+- id: 10
+  title: "Room Context Commands Test"
+  script: '10-context'
+  description: >-
+    This room shows off commands that are only active in a particular room. Try out the <b><cyan>roomtest</cyan></b> command.
+  behaviors:
+    commands: [ 'roomtest' ]
+  items:
+    - id: "limbo:test_key"
+  exits:
+    - roomId: "limbo:8"
+      direction: up
+    - roomId: "limbo:11"
+      direction: east
+  doors:
+    "limbo:8":
+      lockedWith: false
+      closed: true
+- id: 11
+  title: "Locked room with key"
+  description: This room requires a key to get into
+  items:
+    - id: "limbo:locked_chest"
+      respawnChance: 5
+      replaceOnRespawn: true
+  exits:
+    - roomId: "limbo:10"
+      direction: west
+  doors:
+    "limbo:10":
+      lockedBy: "limbo:test_key"
+      locked: true
+      closed: true
