+++ bundles/myelin-areas/areas/limbo/scripts/items/1-shiv.js
@@ -1,0 +1,13 @@
+'use strict';
+
+module.exports = (srcPath) => {
+  const Broadcast = require(srcPath + 'Broadcast');
+
+  return  {
+    listeners: {
+      get: state => function (player) {
+        Broadcast.sayAt(player, `You nearly stab yourself with the pointy end as you pick up ${this.name}`);
+      }
+    }
+  };
+};
