+++ bundles/myelin-areas/areas/limbo/scripts/rooms/8-waypoint.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = (srcPath) => {
+  const Broadcast = require(srcPath + 'Broadcast');
+
+  return  {
+    listeners: {
+      playerEnter: state => function (player) {
+        Broadcast.sayAt(player);
+        Broadcast.sayAt(player, `<b><cyan>Hint: Waypoints allow you to travel vast distances. Save waypoints with '<white>waypoint save</white>', set your preferred home with '<white>waypoint home</white>. If you have enough energy you can return home at any time with '<white>recall</white>'.</cyan></b>`, 80);
+      }
+    }
+  };
+};
