+++ bundles/myelin-areas/areas/mapped/items.yml
@@ -1,0 +1,2 @@
+---
+# Documentation: http://ranviermud.com/extending/areas/items/
