+++ bundles/myelin-areas/areas/mapped/manifest.yml
@@ -1,0 +1,4 @@
+---
+title: Map Test
+info:
+  respawnInterval: 60
