+++ bundles/myelin-areas/areas/mapped/npcs.yml
@@ -1,0 +1,10 @@
+---
+- id: squirrel
+  keywords: ['squirrel']
+  name: A Squirrel
+  level: 2
+  description: "A furry little squirrel"
+  behaviors:
+    ranvier-wander:
+      interval: 30
+      areaRestricted: true
